# openapi.yaml
openapi: 3.1.0
info:
  contact:
    url: https://www.asam.net/project-detail/sovd-service-oriented-vehicle-diagnostics/
  description: Definition of ASAM SOVD API
  license:
    name: Licensing Terms for Grant of Rights to Use ASAM Products
    url: https://www.asam.net/license/
  title: Service-Oriented Vehicle Diagnostic API
  version: 1.0.0
servers:
- url: https://sovd.server/v1
...
paths:
...
  /{entity-collection}/{entity-id}/updates/{update-package-id}:
    delete:
      description: The method deletes an update package from SOVD server. The resource
        is afterwards not available anymore.
      parameters:
      - description: Path to an entity collection
        explode: false
        in: path
        name: entity-collection
        required: true
        schema:
          enum:
          - Area
          - Component
          - App
          - Function
          type: string
        style: simple
      - description: ID of an entity
        explode: false
        in: path
        name: entity-id
        required: true
        schema:
          type: string
        style: simple
      - description: Identifier for the update package
        explode: false
        in: path
        name: update-package-id
        required: true
        schema:
          type: string
        style: simple
      responses:
        "204":
          description: The request was successful.
        "405":
          description: Deleting is not foreseen for an autonomous update-package-id.
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/__any_path__docs_get_default_response'
          description: An unexpected error occurred.
      tags:
      - updates
    get:
      description: The method returns detailed information for a given update package.
      parameters:
      - description: Path to an entity collection
        explode: false
        in: path
        name: entity-collection
        required: true
        schema:
          enum:
          - Area
          - Component
          - App
          - Function
          type: string
        style: simple
      - description: ID of an entity
        explode: false
        in: path
        name: entity-id
        required: true
        schema:
          type: string
        style: simple
      - description: Identifier for the update package
        explode: false
        in: path
        name: update-package-id
        required: true
        schema:
          title: id
          type: string
        style: simple
      - description: "Specifies whether the response should include schema information\
          \ or not. Default: false"
        explode: true
        in: query
        name: include-schema
        required: false
        schema:
          type: boolean
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/__entity_collection___entity_id__updates__update_package_id__get_200_response'
          description: The request was successful.
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/__any_path__docs_get_default_response'
          description: An unexpected error occurred.
      tags:
      - updates
...
  /ecu/engine/updates/{update_id}:
    get:
      description: Check the status of a software update for the engine ECU.
      operationId: getEngineEcuUpdateStatus
      parameters:
      - explode: false
        in: path
        name: update_id
        required: true
        schema:
          title: id
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
          description: Update status
        "404":
          description: Update not found
      summary: Get Update Status
      tags:
      - updates
components:
  schemas:
...
    __entity_collection___entity_id__updates__update_package_id__get_200_response:
      example:
        notes_translation_id: notes_translation_id
        schema: "{}"
        notes: notes
        origin:
        - proximity: proximity
          remote: remote
        - proximity: proximity
          remote: remote
        preconditions_transaltion_id: preconditions_transaltion_id
        user_activity: user_activity
        affected_components:
        - affected_components
        - affected_components
        update_name: update_name
        duration: 0
        automated: true
        execution_conditions: execution_conditions
        size: 6
        preconditions: preconditions
        update_translation_id: update_translation_id
        updated_components:
        - updated_components
        - updated_components
        id: id
        user_activity_translation_id: user_activity_translation_id
      properties:
        id:
          description: "Identifier of the update package. Condition: Only provided\
            \ if the path parameter update-package-id autonomous was used."
          title: id
          type: string
        update_name:
          description: Name of the update
          title: update_name
          type: string
        automated:
          description: "Specifies whether the update package can be installed automatedly\
            \ or not. If missing, the attribute defaults to false."
          title: automated
          type: boolean
        origin:
          description: "List of origins for which the update package is applicable.\
            \ If the attribute is left out or empty, any origin is applicable."
          items:
            $ref: '#/components/schemas/__entity_collection___entity_id__updates_get_origin_parameter'
          title: origin
          type: array
        update_translation_id:
          description: Translation identifier for the update_name.
          title: update_translation_id
          type: string
        notes:
          description: "Release notes for the updates, may also be a URI."
          title: notes
          type: string
        notes_translation_id:
          description: Translation identifier for the notes.
          title: notes_translation_id
          type: string
        user_activity:
          description: Activities which have to be performed by the user after the
            update.
          title: user_activity
          type: string
        user_activity_translation_id:
          description: Translation identifier for the user_activity.
          title: user_activity_translation_id
          type: string
        preconditions:
          description: "Preconditions which shall be fulfilled for installing the\
            \ update (e.g., doors locked, parking, …). In general, not being able\
            \ to fulfill the preconditions would block the preparation of the update."
          title: preconditions
          type: string
        preconditions_transaltion_id:
          description: Translation identifier for the preconditions.
          title: preconditions_transaltion_id
          type: string
        execution_conditions:
          description: Description of the conditions which are required for executing
            the update. May only be set if status is completed.
          title: execution_conditions
          type: string
        duration:
          description: "Time for which the vehicle will not be available when the\
            \ update is installed and activated. The time is defined in seconds. If\
            \ not set, the availability of the vehicle would not be affected."
          title: duration
          type: integer
        size:
          description: The download size of the update. The size is defined in kilo
            bytes.
          title: size
          type: integer
        updated_components:
          description: "A list of Components to be updated. The URIs are in the form\
            \ of {entity-path}."
          items:
            format: uri-reference
            title: href
            type: string
          title: updated_components
          type: array
        affected_components:
          description: "The list documents which Components may have a different behavior\
            \ etc. The focus lies on user-perceived functions and Components. The\
            \ URIs are in the form of {entity-path}."
          items:
            format: uri-reference
            title: href
            type: string
          title: affected_components
          type: array
        schema:
          description: |
            Placeholder schema definition. (External OpenAPI 3.1 schema reference removed for generator compatibility)
          title: schema
          type: object
      required:
      - size
      - update_name
      title: __entity_collection___entity_id__updates__update_package_id__get_200_response
...
    __any_path__docs_get_default_response:
      description: It is recommended that the response body of an error response includes
        a more detailed description of the error which helps in diagnosing the issue.
      example:
        vendor_code: vendor_code
        translation_id: translation_id
        error_code: error_code
        message: message
        parameters: "{}"
      properties:
        error_code:
          description: SOVD standardized error code
          title: error_code
          type: string
        vendor_code:
          description: "Vendor specific error code Condition: If the error_code is\
            \ vendor-specific this attribute has to be filled."
          title: vendor_code
          type: string
        message:
          description: Message describing the problem in more detail.
          title: message
          type: string
        translation_id:
          description: Identifier for translating the message
          title: translation_id
          type: string
        parameters:
          description: Key-value-list which provides additional information on the
            error
          title: parameters
          type: object
      required:
      - error_code
      - message
      title: __any_path__docs_get_default_response
...
    __entity_collection___entity_id__updates_get_origin_parameter:
      description: The retrieval of update packages supports the definition of an
        origin to specify in which environment or by whom the software update is triggered
        by the SOVD client.
      example:
        proximity: proximity
        remote: remote
      properties:
        remote:
          description: "The update is performed over-the-air and no technician intervention\
            \ is possible. Based on the requested version, authorization, network\
            \ connection, etc. the SOVD server (or OEM backend) defines which updates\
            \ are applicable."
          title: remote
          type: string
        proximity:
          description: The update is performed in a workshop and potentially technician
            intervention is possible.
          title: proximity
          type: string
      required:
      - proximity
      - remote
      title: __entity_collection___entity_id__updates_get_origin_parameter
...